+++ bayes_people_tracker/include/bayes_people_tracker/asso_exception.h
