+++ bayes_people_tracker/include/bayes_people_tracker/simple_tracking.h
